Removed placeholder files
Imperfect and perfect MCTS merged, proof of concept removed and replaced with main
