--- Imp_MCTS.py
@@ -1,164 +1,0 @@
-import logging
-import math
-
-import numpy as np
-
-EPS = 1e-8
-
-log = logging.getLogger(__name__)
-
-
-# Kind of hacky version of MCTS for imperfect info games that needs to be standardised
-class Imp_MCTS():
-    """
-    This class handles the MCTS tree for imperfect info games, or at least I bloody well hope so.
-    """
-
-    def __init__(self, game, nnet, player, args):
-        self.game = game
-        self.nnet = nnet
-        self.args = args
-        self.Qsa = {}  # stores Q values for s,a (as defined in the paper)
-        self.Nsa = {}  # stores #times edge s,a was visited
-        self.Ns = {}  # stores #times board s was visited
-        self.Ps = {}  # stores initial policy (returned by neural net)
-
-        self.Es = {}  # stores game.getGameEnded ended for board s
-        self.Vs = {}  # stores game.getValidMoves for board s
-        self.player = player
-
-    def set_num(self, number):
-        self.player = number
-
-    def __call__(self, canonicalBoard, temp=0):
-        return np.random.choice(range(self.game.getActionSize()), p=self.getActionProb(canonicalBoard, temp))
-
-    def getActionProb(self, canonicalBoard, temp=1):
-        """
-        This function performs numMCTSSims simulations of MCTS starting from
-        canonicalBoard.
-
-        Returns:
-            probs: a policy vector where the probability of the ith action is
-                   proportional to Nsa[(s,a)]**(1./temp)
-        """
-        for i in range(self.args.numMCTSSims):
-            board = self.game.getModelBoard(canonicalBoard, self.player)
-            # This isn't going to be valid for all games...
-            next_player = self.game.getNextPlayer(board)
-            self.search(board, next_player, -next_player)
-
-        s = self.game.stringRepresentation(canonicalBoard)
-        # Is counts even the best way to do it?? Why not values??
-        counts = [self.Nsa[(s, a)] if (
-            s, a) in self.Nsa else 0 for a in range(self.game.getActionSize())]
-
-        if temp == 0:
-            if sum(counts) != 0:
-                bestAs = np.array(np.argwhere(counts == np.max(counts))).flatten()
-                bestA = np.random.choice(bestAs)
-            else:
-                valids = self.game.getValidMoves(canonicalBoard, self.player)
-                bestA = np.argmax(self.nnet.predict(canonicalBoard * self.player)[0] * valids)
-            probs = [0] * len(counts)
-            probs[bestA] = 1
-            return probs
-
-        if sum(counts) == 0:
-            # I know, I know, I'm sorry
-            valids = self.game.getValidMoves(canonicalBoard, self.player)
-            probs = self.nnet.predict(canonicalBoard * self.player)[0] * valids
-            prob_sum = np.sum(probs)
-            probs /= prob_sum
-            return probs
-
-        counts = [x ** (1. / temp) for x in counts]
-        counts_sum = float(sum(counts))
-        probs = [x / counts_sum for x in counts]
-        return probs
-
-    def search(self, board, player, prev_player):
-        # TODO: Update notes
-        """
-        This function performs one iteration of MCTS. It is recursively called
-        till a leaf node is found. The action chosen at each node is one that
-        has the maximum upper confidence bound as in the paper.
-
-        Once a leaf node is found, the neural network is called to return an
-        initial policy P and a value v for the state. This value is propagated
-        up the search path. In case the leaf node is a terminal state, the
-        outcome is propagated up the search path. The values of Ns, Nsa, Qsa are
-        updated.
-
-        NOTE: the return values are the negative of the value of the current
-        state. This is done since v is in [-1,1] and if v is the value of a
-        state for the current player, then its value is -v for the other player.
-
-        Returns:
-            v: the negative of the value of the current canonicalBoard
-        """
-        canonicalBoard = self.game.getCanonicalForm(board, player) * player
-        multiplier = player * prev_player  # -1 if different, 1 if same
-
-        s = self.game.stringRepresentation(canonicalBoard)
-
-        if s not in self.Es:
-            self.Es[s] = self.game.getGameEnded(canonicalBoard, 1)
-        if self.Es[s] != 0:
-            # terminal node
-            return self.Es[s] * multiplier
-
-        if s not in self.Ps:
-            # leaf node
-            self.Ps[s], v = self.nnet.predict(canonicalBoard)
-            valids = self.game.getValidMoves(canonicalBoard, 1)
-            self.Ps[s] = self.Ps[s] * valids  # masking invalid moves
-            sum_Ps_s = np.sum(self.Ps[s])
-            if sum_Ps_s > 0:
-                self.Ps[s] /= sum_Ps_s  # renormalize
-            else:
-                # if all valid moves were masked make all valid moves equally probable
-
-                # NB! All valid moves may be masked if either your NNet architecture is insufficient or you've get overfitting or something else.
-                # If you have got dozens or hundreds of these messages you should pay attention to your NNet and/or training process.
-                log.error("All valid moves were masked, doing a workaround.")
-                self.Ps[s] = self.Ps[s] + valids
-                self.Ps[s] /= np.sum(self.Ps[s])
-
-            self.Vs[s] = valids
-            self.Ns[s] = 0
-            return v * multiplier
-
-        valids = self.Vs[s]
-        cur_best = -float('inf')
-        best_act = -1
-
-        # pick the action with the highest upper confidence bound
-        for a in range(self.game.getActionSize()):
-            if valids[a]:
-                if (s, a) in self.Qsa:
-                    u = self.Qsa[(s, a)] + self.args.cpuct * self.Ps[s][a] * math.sqrt(self.Ns[s]) / (
-                            1 + self.Nsa[(s, a)])
-                else:
-                    u = self.args.cpuct * self.Ps[s][a] * math.sqrt(self.Ns[s] + EPS)  # Q = 0 ?
-
-                if u > cur_best:
-                    cur_best = u
-                    best_act = a
-
-        a = best_act
-        next_s, next_player = self.game.getNextState(board, player, a)
-        # next_s = self.game.getCanonicalForm(next_s, next_player)
-
-        v = self.search(next_s, next_player, player)
-
-        if (s, a) in self.Qsa:
-            self.Qsa[(s, a)] = (self.Nsa[(s, a)] * self.Qsa[(s, a)] + v) / (self.Nsa[(s, a)] + 1)
-            self.Nsa[(s, a)] += 1
-
-        else:
-            self.Qsa[(s, a)] = v
-            self.Nsa[(s, a)] = 1
-
-        self.Ns[s] += 1
-        return v * multiplier
